(MODULES-1830) Allow later passenger versions
Updated passenger-conf.erb to set PassengerEnabled On instead of
RailsAutoDetect On for all new versions of passenger.

This includes a scoping fix for the spec tests, as the old tests would
set :facts twice in the outer scope.

This includes removing the puppetlabs-conf.erb template in favor of
using the component confs used on Debian, to avoid the code duplication,
which made me almost miss the second copy of the code to fix.